+++ wheatley/row_generation/__init__.py
@@ -3,7 +3,7 @@
 from .complib_composition_generator import ComplibCompositionGenerator
 from .dixonoids_generator import DixonoidsGenerator
 from .go_and_stop_calling_generator import GoAndStopCallingGenerator
-from .method_place_notation_generator import MethodPlaceNotationGenerator
+from .method_place_notation_generator import MethodPlaceNotationGenerator, generator_from_special_title
 from .place_notation_generator import PlaceNotationGenerator
 from .plain_hunt_generator import PlainHuntGenerator
 from .row_generator import RowGenerator